% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/driver_redis.R
\name{driver_redis}
\alias{driver_redis}
\alias{driver_rlite}
\title{Redis object cache driver}
\usage{
driver_redis(prefix, ...)

driver_rlite(prefix, ...)
}
\arguments{
\item{prefix}{Prefix for keys}

\item{...}{Arguments passed through to \code{RedisAPI::hiredis} or
\code{rrlite::hirlite}.}
}
\description{
Redis object cache driver
}
\author{
Rich FitzJohn
}

