% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/driver_external.R
\name{fetch_hook_download}
\alias{fetch_hook_download}
\alias{fetch_hook_download_fmt}
\alias{fetch_hook_read}
\title{Hook for downloading files}
\usage{
fetch_hook_download(furl, fread)

fetch_hook_download_fmt(fmt, fread)

fetch_hook_read(fpath, fread)
}
\arguments{
\item{furl}{Function to convert \code{key, namespace} into a URL.}

\item{fread}{Function for converting \code{filename} into an R
object.}

\item{fmt}{A \code{\link{sprintf}}-style format that \code{key}
can be substituted into to give a URL.}

\item{fpath}{Function to convert \code{key, namespace} into a filename}
}
\description{
Hook for downloading files as an external resource
}
\seealso{
\code{\link{driver_external}}
}

