% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_external.R
\name{fetch_hook_read}
\alias{fetch_hook_read}
\title{Hook to fetch a resource from a file.}
\usage{
fetch_hook_read(fpath, fread)
}
\arguments{
\item{fpath}{Function to convert \code{key, namespace} into a file path}

\item{fread}{Function for converting \code{filename} into an R pobject}
}
\description{
Hook to fetch a resource from a file, for use with driver_external
}
\examples{
hook <- fetch_hook_read(
    function(key, namespace) paste0(key, ".csv"),
    function(filename) read.csv(filename, stringsAsFactors=FALSE))
}

