% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/driver_external.R
\name{github_release_versions}
\alias{github_release_file}
\alias{github_release_versions}
\title{List github releases}
\usage{
github_release_versions(repo, strip_v = TRUE)

github_release_file(repo, file, add_v = TRUE)
}
\arguments{
\item{repo}{Name of a repo in format username/repository}

\item{strip_v}{Strip the leading "v" from version names? (e.g.,
"v1.0.0" becomes "1.0.0"}

\item{file}{filename within the release}

\item{add_v}{Add the leading "v" to the version name (e.g, "1.0.0"
becomes "v1.0.0").}
}
\description{
Utility function for extracting versions of github releases.
}

