% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storr.R
\name{join_key_namespace}
\alias{join_key_namespace}
\title{Recycle key and namespace}
\usage{
join_key_namespace(key, namespace)
}
\arguments{
\item{key}{A vector of keys}

\item{namespace}{A vector of namespace}
}
\value{
A list with elements \code{n}, \code{key} and \code{namespace}
}
\description{
Utility function for driver authors
}
\details{
This exists to join, predictably, keys and namespaces for
operations like \code{mget}.  Given a vector or scalar for
\code{key} and \code{namespace} we work out what the required
length is and recycle \code{key} and \code{namespace} to the
appropriate length.
}
