% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_rds.R
\name{register_mangler}
\alias{register_mangler}
\title{Register a key mangler}
\usage{
register_mangler(name, encode, decode, overwrite = FALSE)
}
\arguments{
\item{name}{character scalar, name of the mangler}

\item{encode}{function to encode keys. Must have arguments \code{x}
and \code{pad}.}

\item{decode}{function to decode keys. Must have arguments \code{x}
and \code{error}}

\item{overwrite}{logical, whether to overwrite a previously
registered mangler.}
}
\value{
nothing
}
\description{
Define custom functinons for mangling \code{storr_rds()} keys.
}
\details{
\code{storr_rds(mangle_key = "base64")} encodes keys using base64
encoding. This precaution ensures that the names of key files
do not have illegal character names such as ":". However,
base 64 encoding can be slow for some applications,
so you have the option of setting your own custom key mangler.
First, create functions to encode and decode keys and register them
with \code{\link{register_mangler}()}.
Next, set \code{mangle_key} equal to the \code{name}
you gave \code{\link{register_mangler}()}.
}
\examples{
register_mangler(
  "test_mangler",
  encode = function (x, pad) {
    paste0("x_", x)
  },
  decode = function(x, error) {
    substr(x, start = 3, stop = 1e7)
  }
)
s <- storr_rds(tempfile(), mangle_key = "test_mangler")
s$set("a", 1)
s$get("a")
list.files(file.path(s$driver$path, "keys", s$default_namespace))
options(storr_mangler = NULL)
}
