% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{driver_redis_api}
\alias{driver_redis_api}
\alias{storr_redis_api}
\title{Defunct functions}
\usage{
driver_redis_api(...)

storr_redis_api(...)
}
\arguments{
\item{...}{parameters (now all dropped as dots)}
}
\description{
Defunct functions
}
\details{
The redis functions (\code{driver_redis_api} and
\code{storr_redis_api}) have been moved out of this package and
into Redis.  I don't believe anyone is using them at the time of
the move so this is being done fairly abruptly - this is
unfortunate, but necessary to avoid a circular dependency!  The
new functions are simply \code{redux::driver_redis_api} and
\code{redux::storr_redis_api}, along with a helper function
\code{redux::storr_hiredis} which also creates the connection.
}
