% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/storr.R
\name{storr}
\alias{storr}
\title{Object cache}
\usage{
storr(driver, default_namespace = "objects")
}
\arguments{
\item{driver}{A driver object}

\item{default_namespace}{Default namespace to store objects in.
By default \code{"objects"} is used, but this might be useful to
have two diffent \code{storr} objects pointing at the same store,
but storing things in different namespaces.}
}
\description{
Create an object cache
}

