% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_external.R
\name{storr_external}
\alias{storr_external}
\title{Storr that kooks for external resources}
\usage{
storr_external(storage_driver, fetch_hook, default_namespace = "objects")
}
\arguments{
\item{storage_driver}{Another \code{storr} driver to handle the
actual storage.}

\item{fetch_hook}{A function to run to fetch data when a key is
not found in the store.  This function must throw an error (of
any type) if the external resource cannot be resolved.}

\item{default_namespace}{Default namespace (see
\code{\link{storr}})}
}
\description{
storr for fetching external resources.  This does not do a full
cascade (that will be implemented elsewhere) but does a very
simple pattern where if a key cannot be found in the storr we go
out to some external source to find it.
}

