% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_redis_api.R
\name{storr_redis_api}
\alias{driver_redis_api}
\alias{storr_redis_api}
\title{Redis object cache driver}
\usage{
storr_redis_api(prefix, con, default_namespace = "objects")

driver_redis_api(prefix, con)
}
\arguments{
\item{prefix}{Prefix for keys.  We'll generate a number of keys
that start with this string.  Probably terminating the string
with a punctuation character (e.g., ":") will make created
strings nicer to deal with.}

\item{con}{A \code{redis_api} connection object, as created by the
RedisAPI package.  This package does not actually provide the
tools to create a connection; you need to provide one that is
good to go.}

\item{default_namespace}{Default namespace (see \code{\link{storr}}).}
}
\description{
Redis object cache driver
}
\author{
Rich FitzJohn
}

