% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_environment.R
\name{driver_environment}
\alias{driver_environment}
\alias{storr_environment}
\title{Environment object cache driver}
\usage{
driver_environment(envir = new.env(parent = emptyenv()))

storr_environment(default_namespace = "objects", mangle_key = FALSE)
}
\arguments{
\item{envir}{The environment to point the storr at.  The default
creates an new empty environment which is generally the right
choice.}

\item{default_namespace}{Default namespace (see \code{\link{storr}})}

\item{mangle_key}{Mangle key? (see \code{\link{storr}})}
}
\description{
Dummy environment driver that can possibly be used for the internals
}

