% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/driver_environment.R
\name{driver_environment}
\alias{driver_environment}
\alias{storr_environment}
\title{Environment object cache driver}
\usage{
driver_environment()

storr_environment(default_namespace = "objects", mangle_key = FALSE)
}
\arguments{
\item{default_namespace}{Default namespace (see \code{\link{storr}})}

\item{mangle_key}{Mangle key? (see \code{\link{storr}})}
}
\description{
Dummy environment driver that can possibly be used for the internals
}

