% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/driver_environment.R
\name{driver_environment}
\alias{driver_environment}
\title{Environment object cache driver}
\usage{
driver_environment()
}
\description{
Dummy environment driver that can possibly be used for the internals
}

