% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_environment.R
\name{driver_environment}
\alias{driver_environment}
\alias{storr_environment}
\title{Environment object cache driver}
\usage{
driver_environment(envir = NULL)

storr_environment(envir = NULL, default_namespace = "objects",
  mangle_key = FALSE)
}
\arguments{
\item{envir}{The environment to point the storr at.  The default
creates an new empty environment which is generally the right
choice.  However, if you want multiple environment storrs
pointing at the same environment then pass the \code{envir}
argument along.}

\item{default_namespace}{Default namespace (see \code{\link{storr}})}

\item{mangle_key}{Mangle key? (see \code{\link{storr}})}
}
\description{
Fast but transient environment driver
}

