% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/driver_external.R
\name{driver_external}
\alias{driver_external}
\title{Driver for fetching exernal resources}
\usage{
driver_external(storage_driver, fetch_hook)
}
\arguments{
\item{storage_driver}{Another \code{storr} driver to handle the
actual storage.}

\item{fetch_hook}{A function to run to fetch data when a key is
not found in the store.}
}
\description{
storr driver for fetching external resources
}

