% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/driver_rds.R
\name{driver_rds}
\alias{driver_rds}
\title{RDS object cache driver}
\usage{
driver_rds(path)
}
\arguments{
\item{path}{Path for the store.  \code{tempdir()} is a good choice
for ephemeral storage, \code{rappdirs} might be nice for
application data.}
}
\description{
RDS object cache driver
}

