% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/driver_rds.R
\name{driver_rds}
\alias{driver_rds}
\alias{storr_rds}
\title{RDS object cache driver}
\usage{
driver_rds(path, compress = FALSE)

storr_rds(path, compress = FALSE, default_namespace = "objects",
  mangle_key = FALSE)
}
\arguments{
\item{path}{Path for the store.  \code{tempdir()} is a good choice
for ephemeral storage, \code{rappdirs} might be nice for
application data.}

\item{compress}{Compress the generated file?  This saves a small
amount of space for a reasonable amount of time.}

\item{default_namespace}{Default namespace (see \code{\link{storr}})}

\item{mangle_key}{Mangle key? (see \code{\link{storr}})}
}
\description{
RDS object cache driver
}

