% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/driver_redis.R
\name{driver_redis}
\alias{driver_redis}
\alias{driver_redis_api}
\alias{driver_rlite}
\alias{storr_redis}
\alias{storr_redis_api}
\alias{storr_rlite}
\title{Redis object cache driver}
\usage{
driver_redis(prefix, ...)

driver_rlite(prefix, ...)

driver_redis_api(prefix, con)

storr_redis(prefix, ..., default_namespace = "objects", mangle_key = FALSE)

storr_rlite(prefix, ..., default_namespace = "objects", mangle_key = FALSE)

storr_redis_api(prefix, con, default_namespace = "objects",
  mangle_key = FALSE)
}
\arguments{
\item{prefix}{Prefix for keys}

\item{...}{Arguments passed through to \code{RedisAPI::hiredis} or
\code{rrlite::hirlite}.}

\item{con}{A \code{redis_api} object}

\item{default_namespace}{Default namespace (see \code{\link{storr}})}

\item{mangle_key}{Mangle keys? (see  \code{\link{storr}})}
}
\description{
Redis object cache driver
}
\author{
Rich FitzJohn
}

