% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_rds.R
\name{register_mangler}
\alias{register_mangler}
\title{Register a key mangler}
\usage{
register_mangler(name, encode, decode, overwrite = FALSE)
}
\arguments{
\item{name}{character scalar, name of the mangler}

\item{encode}{function to encode keys. Must have arguments \code{x}
and \code{pad}.}

\item{decode}{function to decode keys. Must have arguments \code{x}
and \code{error}}

\item{overwrite}{logical, whether to overwrite a previously
registered mangler.}
}
\value{
nothing
}
\description{
Define custom functinons for mangling \code{storr_rds()} keys.
}
