% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/storr.R
\name{storr}
\alias{storr}
\title{Object cache}
\usage{
storr(driver, default_namespace = "objects", mangle_key = FALSE)
}
\arguments{
\item{driver}{A driver object}

\item{default_namespace}{Default namespace to store objects in.
By default \code{"objects"} is used, but this might be useful to
have two diffent \code{storr} objects pointing at the same store,
but storing things in different namespaces.  Do not use namespaces
beginning with \code{"storr_"} as these are used by storr itself
and your data may get overwritten (unlikely though).}

\item{mangle_key}{Mangle keys?  If TRUE, then the key is run
through a hash function first; this allows storing keys against
names that include characters that might not be supported by the
underlying driver.  At present the hash function mangles the
\emph{string} \code{key} but future versions might allow mangling
the \code{value} of \code{key}.}
}
\description{
Create an object cache
}

