% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_redis_api.R
\name{storr_redis_api}
\alias{driver_redis_api}
\alias{storr_redis_api}
\title{Redis object cache driver}
\usage{
storr_redis_api(prefix, con, hash_algorithm = NULL,
  default_namespace = "objects")

driver_redis_api(prefix, con, hash_algorithm = NULL)
}
\arguments{
\item{prefix}{Prefix for keys.  We'll generate a number of keys
that start with this string.  Probably terminating the string
with a punctuation character (e.g., ":") will make created
strings nicer to deal with.}

\item{con}{A \code{redis_api} connection object, as created by the
redux or rrlite packages.}

\item{hash_algorithm}{Name of the hash algorithm to use.  Possible
values are "md5", "sha1", and others supported by
\code{\link{digest}}.  If not given, then we will default to
"md5".}

\item{default_namespace}{Default namespace (see \code{\link{storr}}).}
}
\description{
Redis object cache driver
}
\author{
Rich FitzJohn
}

